Update for initial startup check
Since the SDR runs on an RPi that shutdowns for a storm event, it may cause a jump in strikes as it starts up after the off period. To keep this from causing another shutdown, the client code looks for a not 100% before continuing.
I'll have to think about this more.